factored compression() out into bin/compress.rs
